Publish binary as dagger-cue
dagger-cue namespace in S3 so that it doesn't get mixed with dagger CLI v0.2.x

Do not publish as a GitHub Release until we have clarified how the other
SDKs will get published. We don't know for sure how we are going to
approach those just yet, we don't want to be back-tracking - not doing
anything in this respect is the best option right now.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>