Export GitHub Actions Workflows traces to Honeycomb.io
The goal is for us to better understand where our workflows spend time, including:
- How long are jobs waiting for available runners?
- How long do steps before & after Dagger take?
- How do we check that Dagger Cloud reporting is accurate?

The last one is especially important to me since it's directly related
to **always run 2 of everything**, where Dagger Cloud is 1/2, and this
would be 2/2 (a check for the check).

Without this addition, it's very difficult to know where our workflows
as a whole spend most time, and therefore where we should focus our next
optimisations.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>